requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile
#
aiocache==0.11.1
    # via -r requirements.in
aioredis==1.3.1
    # via -r requirements.in
async-timeout==3.0.1
    # via aioredis
certifi==2020.12.5
    # via elasticsearch
elasticsearch==7.10.1
    # via -r requirements.in
fluent-logger==0.9.6
    # via -r requirements.in
hiredis==1.1.0
    # via aioredis
iso8601==0.1.13
    # via -r requirements.in
msgpack==0.6.2
    # via fluent-logger
psql-replication==0.0.6
    # via -r requirements.in
psycopg2==2.8.6
    # via psql-replication
pytz==2020.5
    # via -r requirements.in
pyyaml==5.4.1
    # via -r requirements.in
redis==3.5.3
    # via -r requirements.in
urllib3==1.26.2
    # via elasticsearch
